D-FINE_analyzer/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
pycocotools>=2.0.7
timm>=0.9.0
scipy>=1.11.0
PyYAML>=6.0
tqdm>=4.65.0
